# blog/docs/swagger.yaml
openapi: 3.0.3
info:
  description: "This is server for blog-platform server."
  version: "1.0.11"
  title: "Swagger blog-platform"
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/post/{post_id}/create-tag:
    post:
      tags:
        - tag
      summary: "Create New Tag"
      description: "Create New Tag"
      operationId: "CreateTag"
      parameters:
        - name: post_id
          in: path
          description: id of the particular post
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTag'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreateTag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateTag'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '405':
          description: Invalid input
  /api/post/{post_id}/tags/{tag_id}:
    get:
      tags:
        - tag
      summary: Find tag by ID
      description: Returns a single tag
      operationId: getTagById
      parameters:
        - name: post_id
          in: path
          description: ID of particular post
          required: true
          schema:
            type: integer
            format: int64
        - name: tag_id
          in: path
          description: tag id that need to be return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: tag not found
    put:
      tags:
        - tag
      summary: Updates a tag in the store with form data
      description: ''
      operationId: updateTagWithForm
      parameters:
        - name: post_id
          in: path
          description: ID of particular post
          required: true
          schema:
            type: integer
            format: int64
        - name: tag_id
          in: path
          description: ID of tag that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
          application/xml:
            schema:
              $ref: '#/components/schemas/Tag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '200':
          description: successful tag update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          description: Bad Request

    delete:
      tags:
        - tag
      summary: Deletes a tag
      description: delete a tag
      operationId: deleteTag
      parameters:
        - name: post_id
          in: path
          description: ID of particular post
          required: true
          schema:
            type: integer
            format: int64
        - name: tag_id
          in: path
          description: tag id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: No Content
        '400':
          description: Invalid tag value
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: theUser
        created_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
        updated_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
      xml:
        name: user
...
    CreateTag:
      type: object
      properties:
        name:
          type: string
          example: "classic"
      xml:
        name: tag
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "classic"
        post_id:
          type: integer
          example: 1
    TagResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "classic"
        post_id:
          type: integer
          example: 1
        post:
          $ref: '#/components/schemas/Post'
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          example: "classic"
        content:
          type: string
          example: "test blog"
        author:
          $ref: '#/components/schemas/User'
        tags_id:
          type: integer
          example: 1
        author_id:
          type: integer
          example: 1
        tags:
          type: array
          items:
            type: string
        comments:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
        updated_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
...